requirements.txt
Flask>=2.2
Flask-WTF>=1.0
Flask-Login>=0.6
Werkzeug>=2.1
mysql-connector-python>=8.0
